+++ libs/create-app/lib/builtins/app-feature/config.js
@@ -1,8 +1,0 @@
-"use strict";
-
-require("source-map-support/register");
-
-module.exports = {
-  disablePackageLock: true
-};
-//# sourceMappingURL=config.js.map