+++ libs/create-app/lib/index.js
@@ -1,8 +1,0 @@
-"use strict";
-
-require("source-map-support/register");
-
-exports.runner = require("./runner");
-exports.steps = require("./steps");
-exports.packageConfig = require("./steps/packageConfig");
-//# sourceMappingURL=index.js.map