+++ libs/types/cjs/types.js
@@ -117,6 +117,9 @@ class TypeSystem {
     }
     callType(method) {
         return (value, typeInfo, i18n, fieldPath)=>{
+            if (typeInfo.type == null) {
+                throw new _errors.InvalidArgument(`Missing type info: ${JSON.stringify(typeInfo)}`);
+            }
             if (!this.primitives.has(typeInfo.type)) {
                 throw new _errors.InvalidArgument(`Unsupported primitive type: "${typeInfo.type}".`);
             }